BF: Give proper response when MaxUpload param is specified
Currently, the value of `MaxUploads` for listMultipartUpload API is
the number of MPU in a bucket. This fix changes the behavior so
`MaxUpload` value is specified by the user, or else is the default of
1000.

This fix also allows for handling of 0 `MaxUploads`. If `MaxUploads`
is specified as 0, then the behavior of AWS is to return an empty
`Uploads` array and set IsTruncated to `false`. A new test has been
added to verify this behavior.